Fix command line runner functionality
