+++ lib/encrypt.rb
@@ -1,35 +1,3 @@
-require 'pry'
-require './lib/enigma'
-class Encrypt
+require './lib/encrypt_from_file'
 
-  def initialize(input = nil, output = nil)
-    @input = input || ARGV[0]
-    @output = output || ARGV[1]
-    create_and_write_encrypted_file
-  end
-
-
-  def open_and_save_file_contents
-    file_contents = File.open(@input, 'r').read
-    file_contents
-    #binding.pry
-  end
-
-  def encrypt_file_contents
-    original_file_contents = open_and_save_file_contents
-    #binding.pry
-    enigma = Enigma.new
-    enigma.encrypt(original_file_contents)
-    #binding.pry
-  end
-
-  def create_and_write_encrypted_file
-    contents = encrypt_file_contents
-    #binding.pry
-    encrypted_file = File.open(@output, "w")
-    encrypted_file.write(contents)
-    encrypted_file.close
-    #binding.pry
-  end
-
-end
+e = EncryptFromFile.new
