--- test/fixtures/encrypted_nick.txt
@@ -1,1 +1,0 @@
-  onuq
