+++ test/test_runner.rb
@@ -6,4 +6,4 @@ require './test/encryption_generator_test'
 require './test/key_generator_test'
 require './test/offset_generator_test'
 require './test/enigma_test'
-require './test/encrypt_test'
+require './test/encrypt_from_file_test'
